+++ VR_Toy_Project/Assets/_PJK/Scripts/Boss/BossAttack/WeakPoint.cs
@@ -6,7 +6,7 @@ public class WeakPoint : MonoBehaviour, IDamageable
 {
     public Boss boss;
 
-    public float weakpointhp = 100f;
+    public int weakpointhp = 100;
 
     private void Update()
     {
@@ -20,11 +20,11 @@ public class WeakPoint : MonoBehaviour, IDamageable
     {
         // 스케일, 체력 초기화
         transform.localScale = new Vector3(0.1f, 0.1f, 0.1f);
-        weakpointhp = 100f;
+        weakpointhp = 100;
         boss.weakActiveCount--;
     }
 
-    public void OnDamage(float damage)
+    public void OnDamage(int damage)
     {
         weakpointhp -= damage;
     }
